Update events API xapi demo for v1 usage
v1 of events API removed the ability to subscribe directly to on-page
events, which is how this demo was getting the events to show at the top
of the page. We instead have to subscribe to the real event stream
coming down from events. This has the advantage that it's a full
end-to-end demo: it shows that the event is published up to our service,
and then comes down in the `/subscribe` call.